// packages/sensinact-mapping-wizard/src/generated/LocationMode.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * LocationMode
 * @generated
 */
export const LocationMode = {
  NONE: 'NONE',
  STATIC: 'STATIC',
  FROM_DATA: 'FROM_DATA'
} as const;

export type LocationMode = typeof LocationMode[keyof typeof LocationMode];

/**
 * Get enum value by name
 */
export function getLocationMode(name: string): LocationMode | undefined {
  return (LocationMode as Record<string, string>)[name] as LocationMode | undefined;
}

/**
 * Get enum value by literal
 */
export function getLocationModeByLiteral(literal: string): LocationMode | undefined {
  if (literal === 'NONE') return LocationMode.NONE;
  if (literal === 'STATIC') return LocationMode.STATIC;
  if (literal === 'FROM_DATA') return LocationMode.FROM_DATA;
  return undefined;
}